+++ Universim/src/universim/starClasses/ClassMStar.java
@@ -4,7 +4,7 @@ import universim.abstractClasses.Star;
 
 public class ClassMStar extends Star{
 
-	public ClassMStar(int x, int y, int size, int temperature, int mass) {
+	public ClassMStar(double x, double y, double size, int temperature, double mass) {
 		super(x, y, size, temperature, mass);
 	}
